update announcement bar
